+++ tests/ResourceTests/ExpenseResourceTest.php
@@ -44,7 +44,7 @@ class ExpenseResourceTest extends ResourceTest
         return $expense;
     }
 
-    private function ensureProject(): ?int
+    protected function ensureProject(): ?int
     {
         $clientId = $this->config->getAnchor('client_id');
         if (!$clientId) {